EES-6440 WIP create redesign page component
^ Conflicts:
^	src/explore-education-statistics-frontend/src/modules/find-statistics/PublicationReleasePage.tsx
