+++ src/explore-education-statistics-frontend/src/modules/find-statistics/components/ReleasePageTitle.module.scss
@@ -1,0 +1,26 @@
+@import '~govuk-frontend/dist/govuk/base';
+
+.releasePageTitleWrap {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-end;
+  gap: govuk-spacing(4);
+}
+
+.releasePageTitle {
+  flex: 1 1 66%;
+}
+
+.subscribeLink {
+  display: inline-flex;
+  gap: govuk-spacing(2);
+  align-items: center;
+  border: 1px solid $govuk-border-colour;
+  text-decoration: none;
+  padding: govuk-spacing(1) govuk-spacing(2);
+  color: $govuk-link-colour;
+
+  svg {
+    fill: $govuk-text-colour;
+  }
+}
